Tweak rollup config and commonjs require statements for bundling
The minified javascript bundle was failing at runtime because it 
couldn't find the modules I was attempting to dynamically require(). 
Turns out CommonJS expects a leading dot for its modules otherwise 
require() will start its search in node_modules instead of the local 
directory. With this commit, the minified bundle has started working on 
the screeps servers for the first time.