Remove maxCount argument from role constructors.
Determining if we need more of a specific role is no longer as simple as 
checking against a max count in all cases. Removing the argument in 
favor of a needsMoreRecruits method that returns false by default and 
must be overridden by subclasses in order to request more of a that 
subclass's role.