Add recruitment roster to Room.prototype with supporting functions.
This commit introduces a system that allows for initialization calls on 
creeps exactly the tick after they are spawned. 

spawnCreep can be called and return OK multiple times a tick on a single 
spawn. The last succesful call will register with the spawn and start 
the spawning process. I've piggybacked this system with my own 
recruitment roster. The creep name on the last successful spawn call 
will be passed to the room's recruit roster. That roster is kept sorted 
by time remaining to spawn. Every tick the front of the roster is 
checked if those recruits have spawned yet. If they have, remove them, 
and initialize their creeps.