Extend Room prototype with function to find walkable space around a position.
The new function checks each position around the supplied coordinates 
(including the coordinates if the flag is set) for walkable terrain. 
Since plains return zero, we can't use a walkable mask so we have to use 
a blocked terrain mask and flip the result. Any terrain that matches the 
blocked terrain mask will be left out of the final results.