Optimize buffer position finding logic.
This is still not the most optimized solution, but it'll do for now. For 
each adjacent walkable square, we check its surroundings. The squares 
with the most available surroundings are candidates for the buffer 
position. Using PathFinder, we find the shortest path from any spawn 
(TODO: handle multiple spawns?) to the candidate set. The last step in 
that shortest path is our buffer position.

NOTE: This still leads to non-optimal buffer positioning. Diagonal moves 
cost the same as moves in the cardinal directions. As such a path might 
arbitrarily choose one of two positions that is the same walking 
distance from our spawn, but is otherwise further from anything else 
than the other candidate position.