Tweak MineSurveyor to build buffers one at a time, in order by distance
Prior to this commit MineSurveyor would plant buffers for all sources in 
a room simultaneously. A call to `survey(roomName)` will now plant a 
construction site on the nearest source without a buffer only. If the 
buffer is still being built, it simply tries again in the same spot and 
fails.

TODO: I swear this game is fucking with me. The RoomPosition objects I'm 
creating are inheriting from Object (as in pos.prototype === 
Object.prototype). I'm manually fixing the prototype before calling 
`PathFinder.search()`. I need to test this in a clean branch to see if 
my code is messing this up somewhere or if it's a true bug in screeps 
server code.