Introduce RoomRunner that grows each room each tick based on its level.
RoomRunner will be responsible for driving a room in between ticks. For 
now the plan is to run each room based on its controller level. This 
seems to make sense as controller levels largely control the capability 
of a room. This might get tricky with remote mining, and when we start 
managing rooms as blocks or colonies.