Dramatically improve per room recruiting efficiency and fidelity.
recruit() will now only attempt to recruit from a given priority list of 
roles (IN THE CORRECT ORDER FINALLY), and attempt to recruit from the 
front until it succeeds. When it does, it skips the rest of the list. 
We're now able to control recruitment priority per call to recruit().