Absolutely disgusting fix that allows for static AND dynamic imports on one module.
First I try to dynamically require a module. If I can't, I try to 
dynamically require a module + _dynamic. Any module that doesn't want to 
crash attempting this needs to create a new file, with the same name but 
ending in _dynamic.js. This file will statically import the original 
module, create a new instance of it, and export it. 

TODO: Remove this ASAP, and seek redemption for your sins.