Role classes and Recruiter no longer decide role demands.
RoomRunner will now hold some logic containers that decide how many of 
each type of role we need to recruit in a room. For the time being 
recruiter will take a role list and will recruit the first role on the 
list it can. In the future a tasker will need to take care of reading 
recruitment demands, tasking idle pioneers if available, and filling the 
rest of the gap through recruitment.