Add RoomRosters class which will manage a cache of rosters on the heap.
This is the first structure being managed in heap memory. Importing the 
module initializes a cache on the `global` object. The cache is an 
object with no prototype, whose properties are room names. Each room 
name points to a roster for the room. Attempting to access a room's 
roster will cause the roster to be created if it doesn't exist. The 
roster is a map of sets, from roles to a set of creep names.