Add Tasker to be used by RoomRunner in assigning roles to Pioneers.
For now tasker takes a room, and a role and a count for which pioneer 
role is needed and how many respectively. Tasker accesses the 
RoomRosters cache for the latest list of pioneers in the room. Up to the 
count of pioneers is retasked, and the number of pioneers retasked is 
returned.