Add ConstructionQueues cache which tracks a room's construction projects.
ConstructionQueues is a per room cache that keeps a cache of future 
construction projects in order. Surveyors are in charge of filling these 
caches at the RoomRunner's request. Surveyor.survey() returns an object 
that gets passed to Surveyor.planConstruction() later. The contents of 
the project object are arbitrary, so long as the surveyor understands 
how to then unpack a project into construction sites later on. The 
project objects are intended to cache the results of expensive surveying 
logic.