Overhaul Room heap access, and add as property on Room.prototype.
I don't know how the code worked prior to this commit. Checking the 
existence of the room's heap was the responsibility of any object 
accessing room heaps. RoomRosters wasn't even checking correctly. 

RoomsHeap now has a single accessor which takes a room and returns its 
corresponding heap from the global rooms heap. If the heap doesn't 
exist, it is created, assigned to the room's name on the rooms heap and 
passed up.

Heap can now also be accessed through a convenient accessor on 
Room.prototype.