Add Foreman to manage projects proposed by surveyors and build them.
Foreman is a new Room level handler, used by RoomRunner. Foreman exposes 
two methods. The first, survey, is intended to be called when the bot is 
initializing or when a room is newly acquired. survey(room) runs all 
known surveyors on the given room, which populate the rooms 
ConstructionQueues with projects for that room. 

run(room) is a three step process. For now the buildType is hardcoded to 
ROADS to test Builder recruitment. First remove completed plans if there 
are any. If all plans were removed, try to generate new ones and build 
them.