Add prettier node module and integrate it with eslint.
The prettier node module contains the actual prettier code formatter. 
Once the Atom extension is installed, this locally installed module will 
be used to run prettier in the IDE. Expecting this, I'm not adding any 
npm scripts for prettier. Instead testing was done with npx in the 
command line. eslint-config-prettier was also installed to make prettier 
play nice with eslint.