Build and deploy scripts now compile all source code using Typescript.
Nothing should change just yet, the build process is the same and puts 
out the same bundle. But a call to typescript's compiler is added using 
a rollup plugin. The typescript compiler is called, and any errors or 
warnings are output by rollup.

The option to stop the build on error is added but commented.