Split roles (creep behavior) and loadouts (creep metadata).
A Role object held information required by the recruiter to initialize 
and cleanup a creep, as well as logic used when running the creep each 
tick in the overseer. On top of that, the Roles constants were 
overloaded in that they decided both a creep's behavior, and metadata. 
Ugly workarounds have been introduced down the line to get around this 
coupling and overloading, such as the _dynamic.js role files.

Decoupling this logic, as this commit does, will tremendously ease 
future development. Notably _dynamic.js was blocking typescript 
migration.