Tweak tsconfig.
Included tweaks:
- Remove esModuleInterop, I think this is only needed if typescript is 
expected to handle commonjs modules, but since rollup's commonjs plugin 
converts those to ES modules before typescript is called, this is not 
needed.
- However allowSyntheticDefaultImports, an implied corollary of 
esModuleInterop, needs to stay on since the compiler will complain if it 
sees imports of modules that export using module.exports. Maybe this is 
an indication that I still don't competely understand the plugin 
mechanics, but soon the codebase will be all Typescript so it won't 
matter.
-More specific includes globs to match rollup.config.
-Upgrade target ES version.
-remove accidentally left in forceConsistentCasingInFileNames option