Re-add mistakenly removed option in tsConfig.
forceConsistentCasingInFileNames was removed because it was thought to 
be an option turning on warnings of file names not matching the names of 
their defined types.

This is not the case, this flags ensures module references and filenames 
are cased the same way. I'm working with both Ubuntu and Windows so this 
is important.