Convert Foreman.js to Foreman.ts
It occurs to me that I should be writing down some learnings, might as 
well do that here.

This commit was a little tricky simply because Object.keys() always 
returns type string[]. (see 
https://stackoverflow.com/questions/52856496/typescript-object-keys-return-string)

Since these keys are actually `BuildableStructureConstant`s, this was 
causing type errors down the line. This was eventually fixed with a 
refactor that included casting the results of Object.keys() to 
BuildableStructureConstant[].