Massage types into Foreman/Surveyor/ConstructionQueues system.
TSC wouldn't stop complaining about the project argument being passed to 
Surveyor.planConstruction. This argument was intended to be any object, 
so long as the surveyor understands it when it's passed back to build 
the construction plans. ESLint+Typescript was very uncomfortable with 
this.

By explicitly typing surveyors and ConstructionQueues, I was able to 
tell Foreman what specific type of surveyor, and therefore surveyor 
project, it was working with. Also explicitely type ConstructionQueues 
with generics based on the typeof of the BuildType argument. This might 
have been unnecessary as foreman's implied types were faulty on the 
surveyor side, not the construction queue side.