Tweak harvester request so there's not always 4 idle harvesters
The old logic picked the max between the current energy needs (energy 
needed to top off the room) and the cost of a Pioneer. In low energy 
need situations, e.g. after the colony has achieved some momentum, this 
results in 4 harvesters sitting there waiting for their targets to empty 
up. 

This tweak puts the Math.max call behind check of the number of creeps 
in the room. The minimum harvester logic will only apply if there are no 
live creeps left.