Tweak foreman logic for checking construction status of planned work.
Previously this method would filter out a position if it could no longer 
find a construction site of the same type at the right position. Since 
construction site could be removed for a variety of reasons, the new 
logic explicitely checks for the the existance of the structure before 
considering its construction complete.